Ensure latents are converted to float16 for ONNX runtime in SDXLPipeline class to improve compatibility and performance during inference.

+++ pipeline_rt/pipeline.py
@@ -203,8 +203,8 @@ class SDXLPipeline:
             print("\n--- Decoding with ONNX VAE (Debug) ---")
             onnx_vae_start_time = time.time()
 
-            # Prepare latents for ONNX runtime
-            onnx_latents = latents.cpu().numpy()
+            # Prepare latents for ONNX runtime, ensuring float16
+            onnx_latents = latents.cpu().numpy().astype(np.float16)
             
             # Run inference
             input_name = self.onnx_vae.get_inputs()[0].name
